# Efficient iteration and search in HashForLabels and HashWithoutLabels