This brought down the time taken by `HashForLabels` from `23+s` for a single query to `<2s`.

In this PR:
* Don't have a range over the slice to avoid copying of slice. Instead, use traditional loop iteration.
* Use binary search instead of linear search (for this I swapped the loops).
* Reuse byte buffer.