@tomwilkie 2019-06-25
> This brought down the time taken by HashForLabels from 23+s for a single query to <2s.

Can you include the query (anonymised) and some information about why this was taking so long?  Ie how many series this was executing over, how many labels those series had etc.