@codesome 2019-06-25
**Query information:**
`count(metric_name{label1="value1"})` from `1561346880` to `1561368480` at step of `15s`, so `1440` instant evals.

**For each point:**
* The `HashForLabels` function is called for each sample. And for each eval there were between 66k-68k samples in the vector (hence 66k-68k series for each eval), hence 66k-68k calls of `HashForLabels` for each of those 1440 points. (Hence there would be ~1440*67k=96.48M calls to `HashForLabels`)
* The metrics had an average of 13 labels.

> What do our benchmarks show for this?

Will post it soon.