@codesome 2019-06-25
```
benchmark                                                                                                             old ns/op      new ns/op      delta
BenchmarkRangeQuery/expr=a_one,steps=1-8                                                                              19302          19171          -0.68%
BenchmarkRangeQuery/expr=a_one,steps=10-8                                                                             19091          19054          -0.19%
BenchmarkRangeQuery/expr=a_one,steps=100-8                                                                            25830          26069          +0.93%
BenchmarkRangeQuery/expr=a_one,steps=1000-8                                                                           84341          79622          -5.60%
BenchmarkRangeQuery/expr=a_ten,steps=1-8                                                                              78516          74314          -5.35%
BenchmarkRangeQuery/expr=a_ten,steps=10-8                                                                             78951          75130          -4.84%
BenchmarkRangeQuery/expr=a_ten,steps=100-8                                                                            148617         142789         -3.92%
BenchmarkRangeQuery/expr=a_ten,steps=1000-8                                                                           707654         694479         -1.86%
BenchmarkRangeQuery/expr=a_hundred,steps=1-8                                                                          661406         641098         -3.07%
BenchmarkRangeQuery/expr=a_hundred,steps=10-8                                                                         670495         656572         -2.08%
BenchmarkRangeQuery/expr=a_hundred,steps=100-8                                                                        1361892        1312128        -3.65%
BenchmarkRangeQuery/expr=a_hundred,steps=1000-8                                                                       7064966        6730311        -4.74%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1-8                                                                    25667          25079          -2.29%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10-8                                                                   26388          25791          -2.26%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=100-8                                                                  42562          41230          -3.13%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1000-8                                                                 178876         174546         -2.42%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1-8                                                                    94014          92347          -1.77%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10-8                                                                   102379         101647         -0.71%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=100-8                                                                  255785         249171         -2.59%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1000-8                                                                 1562183        1535589        -1.70%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1-8                                                                754836         735571         -2.55%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10-8                                                               842924         815177         -3.29%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=100-8                                                              2336350        2310842        -1.09%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1000-8                                                             15412344       15385738       -0.17%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10000-8                                                                1622682        1588805        -2.09%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10000-8                                                                15949513       16011327       +0.39%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10000-8                                                            161149514      155782485      -3.33%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1-8                                                  914548         906290         -0.90%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=10-8                                                 1389211        1391434        +0.16%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=100-8                                                6314528        6249466        -1.03%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1000-8                                               55242749       55234110       -0.02%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1-8                                                  6773722        6677434        -1.42%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=10-8                                                 11382517       11397250       +0.13%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=100-8                                                60059112       60035523       -0.04%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1000-8                                               548486750      547314249      -0.21%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1-8                                              65615572       65755669       +0.21%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=10-8                                             112052726      111655881      -0.35%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-8                                            600199461      599070762      -0.19%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1000-8                                           5473834085     5467663313     -0.11%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1-8                                                                 722350         703802         -2.57%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=10-8                                                                951967         910387         -4.37%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=100-8                                                               3208546        2945246        -8.21%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1000-8                                                              25711877       23514703       -8.55%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1-8                                                                 6128843        5889727        -3.90%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=10-8                                                                8377102        8050571        -3.90%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=100-8                                                               30830365       28767071       -6.69%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1000-8                                                              264117591      235583345      -10.80%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1-8                                                             59804711       58737193       -1.79%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=10-8                                                            82410883       79440843       -3.60%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=100-8                                                           319704929      283991751      -11.17%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1000-8                                                          2568464452     2399256828     -6.59%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1-8                                                                    690151         687425         -0.39%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=10-8                                                                   783564         777811         -0.73%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=100-8                                                                  1617792        1710893        +5.75%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1000-8                                                                 9876116        10975784       +11.13%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1-8                                                                    5797101        5787460        -0.17%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=10-8                                                                   6587075        6688051        +1.53%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=100-8                                                                  15068182       15920167       +5.65%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1000-8                                                                 97622130       109183058      +11.84%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1-8                                                                57017844       56844440       -0.30%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=10-8                                                               65255966       65154896       -0.15%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=100-8                                                              149024187      158969388      +6.67%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1000-8                                                             971165879      1127233976     +16.07%
BenchmarkRangeQuery/expr=-a_one,steps=1-8                                                                             20851          20537          -1.51%
BenchmarkRangeQuery/expr=-a_one,steps=10-8                                                                            21132          20720          -1.95%
BenchmarkRangeQuery/expr=-a_one,steps=100-8                                                                           28850          28490          -1.25%
BenchmarkRangeQuery/expr=-a_one,steps=1000-8                                                                          89870          88451          -1.58%
BenchmarkRangeQuery/expr=-a_ten,steps=1-8                                                                             85007          82511          -2.94%
BenchmarkRangeQuery/expr=-a_ten,steps=10-8                                                                            85371          83644          -2.02%
BenchmarkRangeQuery/expr=-a_ten,steps=100-8                                                                           158411         153734         -2.95%
BenchmarkRangeQuery/expr=-a_ten,steps=1000-8                                                                          737083         728865         -1.11%
BenchmarkRangeQuery/expr=-a_hundred,steps=1-8                                                                         720905         696121         -3.44%
BenchmarkRangeQuery/expr=-a_hundred,steps=10-8                                                                        713192         704351         -1.24%
BenchmarkRangeQuery/expr=-a_hundred,steps=100-8                                                                       1432011        1408703        -1.63%
BenchmarkRangeQuery/expr=-a_hundred,steps=1000-8                                                                      7352775        7104308        -3.38%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1-8                                                                      35846          34909          -2.61%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10-8                                                                     41176          42044          +2.11%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=100-8                                                                    114555         112984         -1.37%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1000-8                                                                   796475         766952         -3.71%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1-8                                                                      173051         171238         -1.05%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10-8                                                                     226882         222546         -1.91%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=100-8                                                                    865660         843054         -2.61%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1000-8                                                                   6901238        6844032        -0.83%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1-8                                                              1546463        1545148        -0.09%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10-8                                                             2103882        2072233        -1.50%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=100-8                                                            8903383        8750914        -1.71%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1000-8                                                           74206811       73219575       -1.33%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10000-8                                                                  7611805        7460231        -1.99%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10000-8                                                                  69361749       67492442       -2.70%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10000-8                                                          745438242      760361515      +2.00%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1-8                                                     79949          79153          -1.00%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=10-8                                                    83349          82385          -1.16%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=100-8                                                   115654         113753         -1.64%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1000-8                                                  417816         406166         -2.79%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1-8                                                     184584         180847         -2.02%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=10-8                                                    207940         205796         -1.03%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=100-8                                                   521402         517597         -0.73%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1000-8                                                  3468530        3500905        +0.93%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1-8                                             1173597        1148865        -2.11%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=10-8                                            1439980        1423725        -1.13%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=100-8                                           4990672        4983428        -0.15%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1000-8                                          39019348       38800205       -0.56%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1-8                                                      79584          80163          +0.73%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=10-8                                                     85321          83365          -2.29%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=100-8                                                    131247         127392         -2.94%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1000-8                                                   561414         545823         -2.78%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=1-8                                                      191291         188624         -1.39%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=10-8                                                     227141         223729         -1.50%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=100-8                                                    660510         655667         -0.73%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=1000-8                                                   4837277        4813489        -0.49%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1-8                                              1210742        1202370        -0.69%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=10-8                                             1585021        1574692        -0.65%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=100-8                                            6225046        6234908        +0.16%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1000-8                                           51553715       52442102       +1.72%
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=1-8                                                  81518          80101          -1.74%
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=10-8                                                 85474          83207          -2.65%
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=100-8                                                132572         127199         -4.05%
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=1000-8                                               560030         529849         -5.39%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=1-8                                                  184780         183638         -0.62%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=10-8                                                 211012         207945         -1.45%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=100-8                                                518791         522143         +0.65%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=1000-8                                               3527522        3456245        -2.02%
BenchmarkRangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=1-8                                          1173459        1161036        -1.06%
BenchmarkRangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=10-8                                         1434814        1427996        -0.48%
BenchmarkRangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=100-8                                        5076768        4985977        -1.79%
BenchmarkRangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=1000-8                                       39194266       39492967       +0.76%
BenchmarkRangeQuery/expr=abs(a_one),steps=1-8                                                                         24122          23386          -3.05%
…